feat: prisma controllers + some fixes
