+++ types/index.ts
@@ -1,0 +1,4 @@
+export * from './wish';
+export * from './stats';
+export * from './modal';
+export * from './components';
